rename developers media file, dir and path
